@@ -12,17 +12,22 @@ description: A new Flutter project.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
+publish_to: none
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
+  flutter: '>=2.5.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  voyager: ^2.0.0
+  voyager:
+    git:
+      url: https://github.com/eyeem/voyager.git
+      ref: v3.0.1
   voyager_bloc:
     path: ../
-  provider: ^4.0.1
+  provider: ^5.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
