@@ -10,12 +10,13 @@ import 'package:voyager_bloc/voyager_bloc.dart';
 
 import './gen/voyager_gen.dart';
 
-class SimpleBlocDelegate extends BlocDelegate {
-  @override
-  void onEvent(Bloc bloc, Object event) {
-    super.onEvent(bloc, event);
-    print(event);
-  }
+class SimpleBlocDelegate extends BlocObserver {
+
+  // @override
+  // void onEvent(Bloc bloc, Object event) {
+  //   super.onEvent(bloc, event);
+  //   print(event);
+  // }
 
   @override
   void onTransition(Bloc bloc, Transition transition) {
@@ -23,14 +24,14 @@ class SimpleBlocDelegate extends BlocDelegate {
     print(transition);
   }
 
-  @override
-  void onError(Bloc bloc, Object error, StackTrace stacktrace) {
-    super.onError(bloc, error, stacktrace);
-    print(error);
-  }
+  // @override
+  // void onError(Bloc bloc, Object error, StackTrace stacktrace) {
+  //   super.onError(bloc, error, stacktrace);
+  //   print(error);
+  // }
 }
 
-final paths = loadPathsFromString('''
+final paths = loadPathsFromYamlString('''
 '/counter' :
   type: counter
   title: Counter
@@ -41,7 +42,7 @@ final paths = loadPathsFromString('''
 ''');
 
 final plugins = [
-  WidgetPluginBuilder().add<CounterPage>((context) => CounterPage()).build(),
+  WidgetPluginBuilder().add("CounterPage",(context) => CounterPage()).build(),
   BlocsPluginBuilder()
       .addBaseBloc<CounterBloc>(
           (context, config, repo) => CounterBloc.fromConfig(config))
@@ -51,7 +52,8 @@ final plugins = [
 ];
 
 void main() async {
-  BlocSupervisor.delegate = SimpleBlocDelegate();
+  // 5.0.0 BREAKING: Remove BlocSupervisor and rename BlocDelegate to BlocObserver.
+  // BlocSupervisor.delegate = SimpleBlocDelegate();
   runApp(Provider.value(
     value: await loadRouter(paths, plugins),
     child: App(),
@@ -63,7 +65,7 @@ class App extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      home: VoyagerWidget(path: pathCounter),
+      home: Text("")//VoyagerWidget(path: pathCounter),
     );
   }
 }
@@ -72,6 +74,7 @@ class CounterPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final voyager = context.voyager;
+
     // ignore: close_sinks
     final counterBloc = voyager.blocs.find<CounterBloc>();
     // ignore: close_sinks
@@ -103,7 +106,7 @@ class CounterPage extends StatelessWidget {
                 child: FloatingActionButton(
                   child: Icon(Icons.add),
                   onPressed: () {
-                    counterBloc.add(CounterEvent.increment);
+                    counterBloc?.add(CounterEvent.increment);
                   },
                 ),
               ),
@@ -112,7 +115,7 @@ class CounterPage extends StatelessWidget {
                 child: FloatingActionButton(
                   child: Icon(Icons.remove),
                   onPressed: () {
-                    counterBloc.add(CounterEvent.decrement);
+                    counterBloc?.add(CounterEvent.decrement);
                   },
                 ),
               ),
@@ -121,7 +124,7 @@ class CounterPage extends StatelessWidget {
                 child: FloatingActionButton(
                   child: Icon(Icons.update),
                   onPressed: () {
-                    themeBloc.add(ThemeEvent.toggle);
+                    themeBloc?.add(ThemeEvent.toggle);
                   },
                 ),
               ),
@@ -136,16 +139,15 @@ class CounterPage extends StatelessWidget {
 enum CounterEvent { increment, decrement }
 
 class CounterBloc extends Bloc<CounterEvent, int> {
-  CounterBloc(this._initalState) : super();
+  CounterBloc(this.initalState) : super((){
+    return 0;
+  }());
 
   factory CounterBloc.fromConfig(dynamic config) {
     return CounterBloc(int.parse(config.toString()));
   }
 
-  final int _initalState;
-
-  @override
-  int get initialState => _initalState;
+  final int initalState;
 
   @override
   Stream<int> mapEventToState(CounterEvent event) async* {
@@ -163,7 +165,7 @@ class CounterBloc extends Bloc<CounterEvent, int> {
 enum ThemeEvent { toggle }
 
 class ThemeBloc extends Bloc<ThemeEvent, ThemeData> {
-  ThemeBloc(this._intialState) : super();
+  ThemeBloc(this.intialState) : super(ThemeData());
 
   factory ThemeBloc.fromConfig(dynamic config) {
     ThemeData data;
@@ -178,10 +180,7 @@ class ThemeBloc extends Bloc<ThemeEvent, ThemeData> {
     return ThemeBloc(data);
   }
 
-  final ThemeData _intialState;
-
-  @override
-  ThemeData get initialState => _intialState;
+  final ThemeData intialState;
 
   @override
   Stream<ThemeData> mapEventToState(ThemeEvent event) async* {
